Add files via upload
The "first-page.jpg" is the initial page of our product. From there we would render the page you are witnessing in index.html. Since we are just learning more about javascripts and API we could not implement first-page accordingly in our prototype but plan to do so on our final implementation.